Refactor utils functions
