+++ templates/partials/post-card/video-content.php
@@ -1,6 +1,6 @@
 <?php
 
-require_once 'helpers.php';
+require_once 'utils/helpers.php';
 
 /**
  * Шаблон контента карточки публикации видео для страницы 'Популярное'
