+++ templates/partials/post-details/video-content.php
@@ -1,6 +1,6 @@
 <?php
 
-require_once 'helpers.php';
+require_once 'utils/helpers.php';
 
 /**
  * Шаблон контента публикации видео для страницы просмотра публикации
