+++ templates/post-details.php
@@ -1,6 +1,6 @@
 <?php
 
-require_once 'helpers.php';
+require_once 'utils/helpers.php';
 
 /**
  * Шаблон основного контента страницы просмотра публикации.
